libckteec: remove old ta invocation functions
Remove now unused sks_*_shm() and ck_inovke_ta() and friend. All
cryptoki wrapper functions uses ckteec_invoke_*() and ckteec_*_shm()
functions.
