+++ libckteec/CMakeLists.txt
@@ -20,7 +20,6 @@ set (SRC
 	src/ck_debug.c
 	src/ck_helpers.c
 	src/invoke_ta.c
-	src/invoke_ta2.c
 	src/pkcs11_token.c
 	src/serializer.c
 	src/serialize_ck.c
