+++ libckteec/src/pkcs11_token.c
@@ -12,7 +12,6 @@
 
 #include "ck_helpers.h"
 #include "invoke_ta.h"
-#include "invoke_ta2.h"
 #include "local_utils.h"
 #include "pkcs11_token.h"
